executor: split types into dedicated files
